edit contact using js

+++ src/main/java/com/itechart/bukinevi/logic/database/impl/AttachmentDAOUtil.java
@@ -15,6 +15,7 @@ import java.util.List;
 public class AttachmentDAOUtil extends AbstractDAO implements AttachmentDAO {
     private static final Logger LOGGER = LogManager.getLogger(AttachmentDAOUtil.class.getName());
 
+    @Override
     public void addAttachment(Attachment attachment) {
         final int EMPLOYEEID = attachment.getEmployeeID();
         updatePrepareStatement("INSERT INTO attachments(file_name,date_of_load,comment,employee_id) " +
@@ -32,6 +33,7 @@ public class AttachmentDAOUtil extends AbstractDAO implements AttachmentDAO {
         }
     }
 
+    @Override
     public int updateAttachment(Attachment attachment) {
         String query = "UPDATE attachments SET file_name=?,date_of_load=?,comment=? WHERE id=?";
         updatePrepareStatement(query);
@@ -50,8 +52,9 @@ public class AttachmentDAOUtil extends AbstractDAO implements AttachmentDAO {
         return -1;
     }
 
-    public ArrayList<Attachment> getAttachmentList(int ID) {
-        ArrayList<Attachment> attachmentList = new ArrayList<>();
+    @Override
+    public List<Attachment> getAttachmentList(int ID) {
+        List<Attachment> attachmentList = new ArrayList<>();
         String query = "select * from attachments WHERE employee_id  = " + ID;
 
         try {
@@ -63,31 +66,66 @@ public class AttachmentDAOUtil extends AbstractDAO implements AttachmentDAO {
                 attachment.setId(e.getInt(1));
                 attachment.setEmployeeID(e.getInt(2));
                 attachment.setFileName(e.getString(3));
-                attachment.setLoadDate(e.getTimestamp(4).toLocalDateTime());
+                attachment.setLoadDate(String.valueOf(e.getTimestamp(4).toLocalDateTime()));
                 attachment.setComment(e.getString(5));
-                attachment.setSaved(true);
                 attachmentList.add(attachment);
             }
         } catch (SQLException e) {
             LOGGER.error("can't get attachment list ", e);
-        }finally {
+        } finally {
             this.closeStatement("getAttachmentList");
         }
         return attachmentList;
     }
 
-    public void deleteAttachment(final int ATTACHMENTID) {
+    @Override
+    public void deleteAttachments(List<Attachment> attachments) {
+        if(attachments.isEmpty()){
+            return;
+        }
         String deleteSQL = "DELETE FROM attachments WHERE attachments.id = ?";
 
         try {
             preparedStatement = connection.prepareStatement(deleteSQL);
-            preparedStatement.setInt(1, ATTACHMENTID);
-            preparedStatement.executeUpdate();
+            for (Attachment attachment : attachments) {
+                preparedStatement.setInt(1, attachment.getId());
+                preparedStatement.addBatch();
+                LOGGER.info("delete attachment from BD: " + attachment.getId());
+            }
+            preparedStatement.executeBatch();
         } catch (SQLException e) {
             LOGGER.error("can't delete attachment ", e);
         } finally {
-          this.closePreparedStatement("deleteAttachment");
+            this.closePreparedStatement("deleteAttachment");
         }
-     }
+    }
+
+    @Override
+    public void insertOrUpdatePhone(List<Attachment> attachments, final int EMPLOYEEID) {
+        String query = "INSERT INTO attachments (id,employee_id,file_name, date_of_load,comment) " +
+                "VALUES (?,?,?,?,?) " +
+                "ON DUPLICATE KEY " +
+                "UPDATE employee_id = values(employee_id),file_name = values(file_name)," +
+                " date_of_load = values(date_of_load), comment = values(comment)";
+        updatePrepareStatement(query);
+        try {
+            for (Attachment attachment : attachments) {
+                if (!attachment.isDeleted()) {
+                    preparedStatement.setInt(1, attachment.getId());
+                    preparedStatement.setInt(2, EMPLOYEEID);
+                    preparedStatement.setString(3, attachment.getFileName());
+                    preparedStatement.setTimestamp(4, Timestamp.valueOf(attachment.getLoadDate()));
+                    preparedStatement.setString(5, attachment.getComment());
+                    preparedStatement.addBatch();
+                    LOGGER.info(String.format("updateOrInsert attachment to BD id: %d", attachment.getId()));
+                }
+            }
+            preparedStatement.executeBatch();
+        } catch (SQLException e) {
+            LOGGER.error("can't updateOrInsert attachment to BD ", e);
+        } finally {
+            this.closePreparedStatement("insertOrUpdatePhone");
+        }
+    }
 
 }
